fix : post 관리를 HashMap으로 수정

+++ src/main/java/org/sopt/utils/IDGenUtil.java
@@ -1,9 +1,9 @@
 package org.sopt.utils;
 
 public class IDGenUtil {
-	private static int Id =0 ;
+	private static long Id =0 ;
 
-	public static Integer generateId(){
+	public static Long generateId(){
 		return Id++;
 	}
 }
